Fix: Cast sizeof to size_t in min() function calls
- Fixed compiler error at lines 395 and 552 in main.cpp
- Cast sizeof(buff) to size_t to match available parameter type
- Resolves 'no instance of overloaded function min' error

+++ src/main.cpp
@@ -392,7 +392,7 @@ bool downloadAndInstallFirmware(String url) {
     size_t available = stream->available();
     
     if (available) {
-      int c = stream->readBytes(buff, min(available, sizeof(buff)));
+      int c = stream->readBytes(buff, min(available, (size_t)sizeof(buff)));
       
       if (c > 0) {
         if (Update.write(buff, c) != c) {
@@ -549,7 +549,7 @@ bool downloadAndInstallLittleFS(String url) {
     size_t available = stream->available();
     
     if (available) {
-      int c = stream->readBytes(buff, min(available, sizeof(buff)));
+      int c = stream->readBytes(buff, min(available, (size_t)sizeof(buff)));
       
       if (c > 0) {
         if (Update.write(buff, c) != c) {
